+++ src/Model/Subtask.php
@@ -18,6 +18,7 @@ class Subtask
     public function __construct(
         public readonly int $id,
         public readonly string $name,
+        public readonly ?string $type = null,
         public readonly ?int $taskId = null,
         public readonly ?\DateTimeImmutable $dateAdd = null,
         public readonly ?\DateTimeImmutable $dateEditedAt = null,
@@ -48,6 +49,7 @@ class Subtask
         return new self(
             id: (int) ($data['id'] ?? 0),
             name: isset($data['name']) ? (string) $data['name'] : '',
+            type: isset($data['type']) ? (string) $data['type'] : null,
             taskId: isset($data['task_id']) ? (int) $data['task_id'] : null,
             dateAdd: DateTimeParser::parseDateTime($data['date_add'] ?? null),
             dateEditedAt: DateTimeParser::parseDateTime($data['date_edited_at'] ?? null),