+++ src/Model/Task.php
@@ -22,6 +22,7 @@ class Task
     public function __construct(
         public readonly int $id,
         public readonly string $name,
+        public readonly ?string $type = null,
         public readonly ?\DateTimeImmutable $dateAdd = null,
         public readonly ?\DateTimeImmutable $dateEditedAt = null,
         public readonly ?\DateTimeImmutable $dueDate = null,
@@ -67,6 +68,7 @@ class Task
         return new self(
             id: (int) ($data['id'] ?? 0),
             name: isset($data['name']) ? (string) $data['name'] : '',
+            type: isset($data['type']) ? (string) $data['type'] : null,
             dateAdd: DateTimeParser::parseDateTime($data['date_add'] ?? null),
             dateEditedAt: DateTimeParser::parseDateTime($data['date_edited_at'] ?? null),
             dueDate: DateTimeParser::parseDateTime($data['due_date'] ?? null),